增加理想漫反射材质、统一浮点数精度

+++ Camera/PerspectiveCamera.cpp
@@ -1,23 +1,23 @@
 
 #include "PerspectiveCamera.h"
 
-PerspectiveCamera::PerspectiveCamera(Point3f look_from, Point3f look_at, Vector3f up, float fov, float aspect_ratio) {
-    float theta = Degrees2Radians(fov);
-    float h = std::tan(theta / 2);
+PerspectiveCamera::PerspectiveCamera(Point3d look_from, Point3d look_at, Vector3d up, double fov, double aspect_ratio) {
+    double theta = Degrees2Radians(fov);
+    double h = std::tan(theta / 2);
 
-    float viewport_height = 2.0 * h;
-    float viewport_width = aspect_ratio * viewport_height;
+    double viewport_height = 2 * h;
+    double viewport_width = aspect_ratio * viewport_height;
 
-    Vector3f w = Normalize(look_from - look_at);
-    Vector3f u = Normalize(Cross(up, w));
-    Vector3f v = Cross(w, u);
+    Vector3d w = Normalize(look_from - look_at);
+    Vector3d u = Normalize(Cross(up, w));
+    Vector3d v = Cross(w, u);
 
     origin = look_from;
     horizontal = viewport_width * u;
     vertical = viewport_height * v;
-    start = origin - horizontal / 2.f - vertical / 2.f - w;
+    start = origin - horizontal / 2.0 - vertical / 2.0 - w;
 }
 
-Ray PerspectiveCamera::GetRay(const float &s, const float &t) const {
+Ray PerspectiveCamera::GetRay(const double &s, const double &t) const {
     return Ray(origin, start + s * horizontal + t * vertical - origin);
 }
