+++ Common/Camera.h
@@ -6,6 +6,6 @@
 #include "Vector.h"
 
 struct Camera {
-    virtual Ray GetRay(const float &s, const float &t) const = 0;
+    virtual Ray GetRay(const double &s, const double &t) const = 0;
 };
