+++ Common/Integrator.h
@@ -6,7 +6,7 @@
 #include "../Renderer/Scene.h"
 
 struct Integrator {
-    virtual Color3f Li(const Ray &ray, std::shared_ptr<Scene> scene) = 0;
+    virtual Color3d Li(const Ray &ray, std::shared_ptr<Scene> scene) = 0;
 };
