+++ Integrator/NormalIntegrator.cpp
@@ -1,13 +1,13 @@
 
 #include "NormalIntegrator.h"
 
-inline Color3f NormalIntegrator::Li(const Ray &ray, std::shared_ptr<Scene> scene) {
+inline Color3d NormalIntegrator::Li(const Ray &ray, std::shared_ptr<Scene> scene) {
     HitResult result;
     if (scene->Intersect(ray, result)) {
-        return 0.5f * (result.normal + Color3f(1.f));
+        return 0.5 * (result.normal + Color3d(1.0));
     }
-    float t = 0.5f * (ray.direction.y + 1.0);
-    return (1.0f - t) * Color3f(1.f, 1.f, 1.f) + t * Color3f(0.5f, 0.7f, 1.f);
+    double t = 0.5 * (ray.direction.y + 1.0);
+    return (1.0 - t) * Color3d(1.0, 1.0, 1.0) + t * Color3d(0.5, 0.7, 1.0);
 }
